Broke the code into several more methods.
Added a handle_timer method for epoll.

Added a method to close the client fd.

Added a method to close the pty fd.

Fixed a bug where the client socket would be left open in a weird "can't identify protocol" state after the shutdown. Needed to make sure there was a close accompanying the shutdown of IO on the socket.

Added further documentation.