Split up the transfer data method and bug fixes.
Added a handle_unwritten and handle_normal data methods. The transfer_data method is now a delegator in how data should be processed and where it should go.

Fixed a bug where the server would segfault when a write errored out.

Changed the pwrite test to cat a file which doesn't contain words easily searched (WARN/Unwritten) to detect a partial write.